refactor: 원시 타입을 RandomNumber 객체로 변경 및 테스트에서 TestRandomNumber 사용하도록 변경
