+++ src/test/java/racinggame/domain/vo/TestRandomNumber.java
@@ -1,0 +1,14 @@
+package racinggame.domain.vo;
+
+public class TestRandomNumber extends RandomNumber {
+    private final int value;
+
+    public TestRandomNumber(int number) {
+        this.value = number;
+    }
+
+    @Override
+    public int getValue() {
+        return this.value;
+    }
+}
